Add files via upload
After Covid-19, the demand for digital processes has skyrocketed. The pillars of this digital revolution were the Public Cloud and other Cloud Computing services. Almost every major country was put on lockdown, making it essential to migrate to the cloud to ensure a stable working environment. 

The public cloud provided a solution for businesses that lacked the financial resources to purchase a dedicated cloud service by allowing them access to shared cloud resources with all associated benefits. 

Learn more about the public cloud, its forms, and its benefits in this presentation.